# Use a HashSet for detecting orphan pages